Improve error handling.
Previously, the client middleware which deserializes the body of
error responses to create a descriptive error object would fail if
it didn't understand the content type, or if the body failed to
deserialize to the appropriate error type. The resulting error was
the error raised during the attempt at interpreting the body, such
as "Unsupported content type", and not the error returned by the
request.

This worked poorly even with our own APIs, which return a helpful
HTML document when a route is incorrectly named or misspelled. Since
HTML cannot be deserialized to an error type, the net middleware would
unhelpfully replace this descriptive error information with "Unsupported
content type".

With this change, the middleware will try to convert the response body
to an error, but if it fails it will fall back to just including the
response body in a newly created error message using `catch_all`,
which means HTML errors and the like will not be lost in this middleware.